[REFACTOR] model.py 각 도메인으로 이동, video 별도 도메인으로 분리, comment 모델 sqlmodel로 수정
