+++ domain/channel/repository/channel_repository.py
@@ -1,4 +1,4 @@
-from core.database.model.channel import Channel
+from domain.channel.model.channel import Channel
 from core.database.repository.crud_repository import CRUDRepository
