+++ domain/content_chunk/model/content_chunk.py
