+++ domain/task/repository/task_repository.py
@@ -1,4 +1,4 @@
-from core.database.model.task import Task
+from domain.task.model.task import Task
 from core.database.repository.crud_repository import CRUDRepository
